# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: EV Charging Station API
  description: |
    API για τη διαχείριση σταθμών φόρτισης ηλεκτρικών οχημάτων (EV Charging Stations).
    
    ## Αυθεντικοποίηση
    Η πλειοψηφία των endpoints απαιτεί αυθεντικοποίηση μέσω JWT token.
    Το token λαμβάνεται από το endpoint `/api/v1/auth/signin` και πρέπει να 
    συμπεριλαμβάνεται στο header `Authorization` ως `Bearer <token>`.
    
    ## Ρόλοι Χρηστών
    - **USER**: Βασικός χρήστης - μπορεί να δει φορτιστές, να κάνει κρατήσεις, κ.λπ.
    - **ADMIN**: Διαχειριστής - πλήρης πρόσβαση σε όλες τις λειτουργίες διαχείρισης
  version: 1.0.0
  contact:
    name: softeng25-02 Team
servers:
  - url: http://localhost:9876
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token που λαμβάνεται από το signin endpoint
  schemas:
...
    Point:
      type: object
      properties:
        pointid:
          type: integer
        providerName:
          type: string
        name:
          type: string
        address:
          type: string
        connectorType:
          type: string
        lon:
          type: string
        lat:
          type: string
        status:
          type: string
          enum: [available, in_use, charging, reserved, outage]
        cap:
          type: number
          description: Μέγιστη ισχύς σε kW
        kwhprice:
          type: number
        reserved_by_me:
          type: boolean
        reservationendtime:
          type: string
          nullable: true
          description: Ώρα λήξης κράτησης (YYYY-MM-DD HH:MM)
    Session:
      type: object
      properties:
        starttime:
          type: string
          description: "Format: YYYY-MM-DD HH:MM"
        endtime:
          type: string
        startsoc:
          type: integer
        endsoc:
          type: integer
        totalkwh:
          type: number
        kwhprice:
          type: number
        amount:
          type: number
...
paths:
...
  /api/v1/points:
    get:
      tags: [Points]
      summary: Λήψη όλων των σημείων φόρτισης
      description: |
        Εναλλακτικό API για φορτιστές. Υποστηρίζει φιλτράρισμα κατά status.
        Αν ο χρήστης είναι συνδεδεμένος, επιστρέφεται και η πληροφορία αν
        ο φορτιστής είναι κρατημένος από αυτόν.
      parameters:
        - name: status
          in: query
          description: |
            Φίλτρο status (available, in_use, charging, reserved, outage, outoforder)
          schema:
            type: string
      responses:
        '200':
          description: Λίστα σημείων φόρτισης
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Point'
  /api/v1/points/{id}:
    get:
      tags: [Points]
      summary: Λήψη συγκεκριμένου σημείου φόρτισης
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Πληροφορίες σημείου φόρτισης
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Point'
        '404':
          description: Σημείο δεν βρέθηκε

  # ==================== RESERVATIONS ====================
...
  /api/v1/newsession:
    post:
      tags: [Sessions]
      summary: Δημιουργία νέας συνεδρίας φόρτισης
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pointid, starttime, endtime, totalkwh, kwhprice, amount]
              properties:
                pointid:
                  type: integer
                starttime:
                  type: string
                  format: date-time
                endtime:
                  type: string
                  format: date-time
                startsoc:
                  type: integer
                endsoc:
                  type: integer
                totalkwh:
                  type: number
                kwhprice:
                  type: number
                amount:
                  type: number
      responses:
        '200':
          description: Συνεδρία δημιουργήθηκε
        '400':
          description: Λείπουν απαιτούμενα πεδία ή μη έγκυρος φορτιστής
        '403':
          description: Φορτιστής σε χρήση ή εκτός λειτουργίας
  /api/v1/sessions/{id}/{from}/{to}:
    get:
      tags: [Sessions]
      summary: Λήψη συνεδριών φορτιστή
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID φορτιστή
          schema:
            type: integer
        - name: from
          in: path
          required: true
          description: Ημερομηνία έναρξης (YYYYMMDD)
          schema:
            type: string
          example: "20260101"
        - name: to
          in: path
          required: true
          description: Ημερομηνία λήξης (YYYYMMDD)
          schema:
            type: string
          example: "20260131"
      responses:
        '200':
          description: Λίστα συνεδριών
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '400':
          description: Μη έγκυρο ID ή μορφή ημερομηνίας

  # ==================== POINT STATUS ====================
...